Disambiguate trauma savant labels.
This gives each of the new "variants" of trauma savant a unique suffix
in their labels, so they can more easily be distinguished on a pawn's
health tab, the hediff's info card, and in debug actions, such as
selecting a hediff to give to a pawn.

The player is already able to inspect the hediffs to determine how
they affect the pawn, so indicating there are different variants does
not give them new information, aside from potentially telling them how
many variants there are.